feat(auth): move security settings into identity hub tab bar (#731)
- Add /auth/security route with the security settings page
- Add Security tab to IdentityTabBar, visible only for actor scope identities
- Add showSecurity prop to IdentityTabBar; layout.tsx sets it true when effectiveIdentity.scope === 'actor'
- Update /auth/settings for actor scope: remove old security link, show message to use tabs
- Keep /auth/settings/security as a redirect to /auth/security for backwards compat
