+++ apps/kernel/app/auth/settings/page.tsx
@@ -3,7 +3,6 @@ import { verifySessionToken, getSessionCookieOptions } from '@/src/lib/auth/jwt'
 import { db, identities, identityMembers } from '@/src/db';
 import { eq, and, isNull } from 'drizzle-orm';
 import { redirect } from 'next/navigation';
-import Link from 'next/link';
 import IdentitySettingsPanel from '../components/IdentitySettingsPanel';
 
 export default async function SettingsPage() {
@@ -30,20 +29,11 @@ export default async function SettingsPage() {
     .where(eq(identities.id, effectiveDid))
     .limit(1);
 
-  // Actor scope: show link to security settings
+  // Actor scope: security is now accessible via the Security tab
   if (!identity || identity.scope === 'actor') {
     return (
-      <div className="space-y-4">
-        <div className="bg-zinc-900 border border-zinc-800 rounded-xl p-6">
-          <h2 className="text-sm font-semibold text-white mb-1">Security</h2>
-          <p className="text-sm text-zinc-400 mb-4">Manage your account security settings.</p>
-          <Link
-            href="/auth/settings/security"
-            className="inline-block px-4 py-2 bg-zinc-800 hover:bg-zinc-700 border border-zinc-700 text-zinc-300 text-sm rounded-lg transition-colors"
-          >
-            Security settings →
-          </Link>
-        </div>
+      <div className="text-zinc-500 text-sm py-8">
+        Account settings are available via the tabs above.
       </div>
     );
   }
